fix rubocop errors

+++ app/models/library_entry.rb
@@ -203,21 +203,25 @@ class LibraryEntry < ApplicationRecord
   end
 
   # TODO: will rename this if I think of a better one
-  after_commit :create_or_update_library_entry, on: [:create, :update]
+  after_commit :create_or_update_library_entry, on: %i[create update]
 
   def create_or_update_library_entry
-    MyAnimeListSyncWorker.perform_async(
-      library_entry_id: id,
-      method: 'create/update'
-    ) if sync_to_mal?
+    if sync_to_mal?
+      MyAnimeListSyncWorker.perform_async(
+        library_entry_id: id,
+        method: 'create/update'
+      )
+    end
   end
 
   after_destroy do
-    MyAnimeListSyncWorker.perform_async(
-      user_id: user_id,
-      media_id: media_id,
-      media_type: media_type,
-      method: 'delete'
-    ) if sync_to_mal?
+    if sync_to_mal?
+      MyAnimeListSyncWorker.perform_async(
+        user_id: user_id,
+        media_id: media_id,
+        media_type: media_type,
+        method: 'delete'
+      )
+    end
   end
 end
